[Refactor] 🍎 길찾기 코드 리팩토링
